apps/agent/pyproject.toml
[project]
name = "sync-hire-agent"
version = "0.2.0"
description = "AI Interview Agent for SyncHire powered by Vision-Agents"
requires-python = ">=3.11"
dependencies = [
    # Vision-Agents from GitHub main with Gemini and OpenAI support
    "vision-agents[gemini,getstream,openai]>=0.2.0",
    # Vision-Agents profiling dependency (required by GitHub version)
    "pyinstrument>=5.0.0",
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    # Environment and HTTP
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    # Type hints
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]
